refactor: split aces_processor planner into a package
planner.py (836 lines) exceeded the 500-line cap. Split it into a
planner/ package (resources, ordering, capability_domains,
manifest_validation, operations, core) behind an API-stable __init__
facade re-exporting plan, realization_disclosure, snapshot_delete_order.

Pure move with no behavior change: every external import site and the
aces._compat re-export shim are unchanged, and the full nox verify suite
(--skip-requirement) is green. Removes the planner.py oversized-allowlist
entry and updates durable spec-coverage digests and path references.

Refs #44
