+++ docs/research/specification-coverage/analysis-v1.json
@@ -2,7 +2,7 @@
   "analysis_id": "aces-standardized-specification-coverage-analysis-v1",
   "protocol_revision": "1.0.0",
   "snapshot_id": "aces-standardized-specification-coverage-8bf12ee-v1",
-  "snapshot_sha256": "4f335d337a36e9cc12b71df9b63fc54b8203eebc78c870662feea360c576b926",
+  "snapshot_sha256": "d3785b35f3396c955cfff19f894d2106155fb06c472425a12ae0719bad2e8c47",
   "generated_at": "2026-07-17",
   "execution_status": "complete",
   "classification_counts": {
