+++ specs/formal/planner/dependency-ordering.md
@@ -62,7 +62,7 @@ change.
 
 - shared semantic source of truth: `implementations/python/packages/aces_processor/semantics/planner.py`
 - planner use sites:
-  - `implementations/python/packages/aces_processor/planner.py`
+  - `implementations/python/packages/aces_processor/planner/`
   - `implementations/python/packages/aces_processor/manager.py`
 - property and agreement tests:
   - `implementations/python/tests/test_semantics_planner.py`
